First commited as before_61921_rev1

+++ src/main/java/com/couchbase/client/java/env/DefaultCouchbaseEnvironment.java
@@ -471,7 +471,7 @@ public class DefaultCouchbaseEnvironment extends DefaultCoreEnvironment implemen
         }
 
         @Override
-        public Builder dcpConnectionBufferAckThreshold(double dcpConnectionBufferAckThreshold) {
+        public Builder dcpConnectionBufferAckThreshold(int dcpConnectionBufferAckThreshold) {
             super.dcpConnectionBufferAckThreshold(dcpConnectionBufferAckThreshold);
             return this;
         }
